Refactored new Swift unit tests for UserDescription into their own test class
