+++ NimbleTests/AsynchronousTest.swift
@@ -57,28 +57,4 @@ class AsyncTest: XCTestCase {
             }
         }
     }
-    
-    func testToEventuallyMatch_CustomFailureMessage() {
-        failsWithErrorMessage(
-            "These aren't eventually equal!\n" +
-            "expected to eventually equal <1>, got <0>") {
-            expect { 0 }.toEventually(equal(1), description: "These aren't eventually equal!")
-        }
-    }
-    
-    func testToEventuallyNotMatch_CustomFailureMessage() {
-        failsWithErrorMessage(
-            "These are eventually equal!\n" +
-            "expected to eventually not equal <1>, got <1>") {
-                expect { 1 }.toEventuallyNot(equal(1), description: "These are eventually equal!")
-        }
-    }
-    
-    func testToNotEventuallyMatch_CustomFailureMessage() {
-        failsWithErrorMessage(
-            "These are eventually equal!\n" +
-            "expected to eventually not equal <1>, got <1>") {
-                expect { 1 }.toEventuallyNot(equal(1), description: "These are eventually equal!")
-        }
-    }
 }
