+++ NimbleTests/UserDescriptionTest.swift
@@ -1,0 +1,54 @@
+import XCTest
+import Nimble
+
+class UserDescriptionTest: XCTestCase {
+    
+    func testToMatcher_CustomFailureMessage() {
+        failsWithErrorMessage(
+            "These aren't equal!\n" +
+            "expected to match, got <1>") {
+                expect(1).to(MatcherFunc { expr, failure in false }, description: "These aren't equal!")
+        }
+    }
+    
+    func testNotToMatcher_CustomFailureMessage() {
+        failsWithErrorMessage(
+            "These aren't equal!\n" +
+            "expected to not match, got <1>") {
+                expect(1).notTo(MatcherFunc { expr, failure in true }, description: "These aren't equal!")
+        }
+    }
+    
+    func testToNotMatcher_CustomFailureMessage() {
+        failsWithErrorMessage(
+            "These aren't equal!\n" +
+            "expected to not match, got <1>") {
+                expect(1).toNot(MatcherFunc { expr, failure in true }, description: "These aren't equal!")
+        }
+    }
+    
+    func testToEventuallyMatch_CustomFailureMessage() {
+        failsWithErrorMessage(
+            "These aren't eventually equal!\n" +
+            "expected to eventually equal <1>, got <0>") {
+                expect { 0 }.toEventually(equal(1), description: "These aren't eventually equal!")
+        }
+    }
+    
+    func testToEventuallyNotMatch_CustomFailureMessage() {
+        failsWithErrorMessage(
+            "These are eventually equal!\n" +
+            "expected to eventually not equal <1>, got <1>") {
+                expect { 1 }.toEventuallyNot(equal(1), description: "These are eventually equal!")
+        }
+    }
+    
+    func testToNotEventuallyMatch_CustomFailureMessage() {
+        failsWithErrorMessage(
+            "These are eventually equal!\n" +
+            "expected to eventually not equal <1>, got <1>") {
+                expect { 1 }.toEventuallyNot(equal(1), description: "These are eventually equal!")
+        }
+    }
+
+}
